huge cleanup and conversion of most string handling (especially paths) to unicode please note: format strings must be %hs for char* arguments and %ls for wchar_t*
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@7161 3db68df2-c116-0410-a063-a993310a9797

+++ source/collada/XMLFix.cpp
@@ -166,7 +166,7 @@ void FixBrokenXML(const char* text, const char** out, size_t* outSize)
 	Log(LOG_INFO, "Running FixBrokenXML");
 
 	size_t textSize = strlen(text);
-	xmlDocPtr doc = xmlParseMemory(text, textSize);
+	xmlDocPtr doc = xmlParseMemory(text, (int)textSize);
 
 	xmlNode* root = xmlDocGetRootElement(doc);
 	if (root && processDocument(root))
